Split precision module
